dnsapi/dns_inwx.sh: instead of a generic query, look for our specific domain
I have encountered the problem, that nameserver.list only returns up to 20 domains per default.

Using the domain parameter, we can filter the results on the serverside (and reduce amount of data to be sent).
A drawback of this setup are the additional api calls for sub domains.

Another possible solution would be to use the parameters pagelimit and/or page..

(https://www.inwx.at/en/help/apidoc/f/ch02s09.html#nameserver.list)